requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
apify-client==1.7.1
requests==2.31.0
Werkzeug==2.3.7
reportlab==4.0.4
pandas==2.0.3
bcrypt==4.0.1
SQLAlchemy==2.0.21
Jinja2==3.1.2
gunicorn==21.2.0
